Add compatibility with newer libmicrohttpd access handler types.
Starting with libmicrohttpd 0.9.71, the access handler callback function
type has changed to return an `enum MHD_Result` instead of an `int`, so the
return type of promhttp_handler needs to be corrected based on the version of
libmicrohttpd it's compiled against.

Additionally, the visibility of this function is changed to static, as it's not
exposed by the promhttp.h header file.